Move BotService to Services folder
